+++ src/js/renderers/flash.js
@@ -289,7 +289,12 @@ const FlashMediaElementRenderer = {
 
 		const
 			autoplay = mediaElement.originalNode.autoplay,
-			flashVars = [`uid=${flash.id}`, `autoplay=${autoplay}`, `allowScriptAccess=${flash.options.shimScriptAccess}`],
+			flashVars = [
+				`uid=${flash.id}`,
+				`autoplay=${autoplay}`,
+				`allowScriptAccess=${flash.options.shimScriptAccess}`,
+				`preload=${mediaElement.originalNode.getAttribute('preload') || ''}`
+			],
 			isVideo = mediaElement.originalNode !== null && mediaElement.originalNode.tagName.toLowerCase() === 'video',
 			flashHeight = (isVideo) ? mediaElement.originalNode.height : 1,
 			flashWidth = (isVideo) ? mediaElement.originalNode.width : 1;
@@ -304,10 +309,7 @@ const FlashMediaElementRenderer = {
 		}
 
 		mediaElement.appendChild(flash.flashWrapper);
-
-		if (mediaElement.originalNode !== null) {
-			mediaElement.originalNode.style.display = 'none';
-		}
+		mediaElement.originalNode.style.display = 'none';
 
 		let settings = [];