+++ changelog.md
@@ -8,6 +8,7 @@
 * Fixed typo when setting default player caused issues on Firefox when extending `MediaElementPlayer` object @rafa8626
 * Fixed issue related to iframe renderers not resizing properly after using fullscreen @rafa8626
 * Updated RTMP and Dailymotion sources @rafa8626
+* Added `useDefaultControls` configuration to simplify list of `features` (https://github.com/mediaelement/mediaelement-plugins/issues/74) @rafa8626
 
 *4.2.3 (2017/07/22)*